Move Cilium to ApplicationDefinition YAML
