+++ pkg/applicationdefinitions/application_catalog.go
@@ -24,7 +24,6 @@ import (
 
 	appskubermaticv1 "k8c.io/kubermatic/sdk/v2/apis/apps.kubermatic/v1"
 	kubermaticv1 "k8c.io/kubermatic/sdk/v2/apis/kubermatic/v1"
-	"k8c.io/kubermatic/v2/pkg/defaulting"
 	"k8c.io/kubermatic/v2/pkg/kubernetes"
 	kkpreconciling "k8c.io/kubermatic/v2/pkg/resources/reconciling"
 	"k8c.io/kubermatic/v2/pkg/resources/registry"
@@ -150,8 +149,6 @@ func updateApplicationDefinition(appDef *appskubermaticv1.ApplicationDefinition,
 	for i := range appDef.Spec.Versions {
 		if sysAppConfig.HelmRepository != "" {
 			appDef.Spec.Versions[i].Template.Source.Helm.URL = registry.ToOCIURL(sysAppConfig.HelmRepository)
-		} else {
-			appDef.Spec.Versions[i].Template.Source.Helm.URL = defaulting.SystemApplicationsHelmRepositoryFromConfig(&sysAppConfig)
 		}
 
 		if credentials != nil {
