+++ pkg/applicationdefinitions/system-applications/cluster-autoscaler.yaml
